Move SSM helper class into new callables file.
